+++ packages/frontend/src/components/message/MessageMarkdown.tsx
@@ -1,6 +1,5 @@
 import React, { useContext } from 'react'
 import {
-  parse_desktop_set,
   parse_text,
   ParsedElement,
 } from '@deltachat/message_parser_wasm/message_parser_wasm'
@@ -20,14 +19,12 @@ import { ChatView } from '../../contexts/ChatContext'
 const log = getLogger('renderer/message-markdown')
 
 let parseMessage: (message: string) => ParsedElement[] = m =>
-  parse_desktop_set(m)
+  parse_text(m, false)
 
 SettingsStoreInstance.subscribe(newState => {
-  if (newState?.desktopSettings.experimentalEnableMarkdownInMessages) {
-    parseMessage = m => parse_text(m, true)
-  } else {
-    parseMessage = m => parse_desktop_set(m)
-  }
+  const markDownEnabled =
+    !!newState?.desktopSettings.experimentalEnableMarkdownInMessages
+  parseMessage = m => parse_text(m, markDownEnabled)
 })
 
 function renderElement(